Move libcurl/libuv HTTP client into acyclic http_client component
Move src/http/curl.h to src/http_client/curl.h, and src/http/test/curl_test.cpp
to src/http_client/test/curl_test.cpp, updating all in-tree includes and the
CMake registration. No forwarding header is left at the old path since curl.h
is an internal header.

Narrow the http component's declared internal dependencies to
[ccf-api, crypto, ds] now that the libuv-dependent client code has moved out,
and add the new http_client component with [ccf-api, ds, uv].

This keeps http protocol-focused and acyclic, and keeps the new http_client
component acyclic and outside the cyclic core.
