+++ scripts/source-dependencies.json
@@ -34,6 +34,7 @@
     "tls": ["crypto", "ds"],
     "udp": ["ds"],
     "uv": [],
-    "http": ["ccf-api", "crypto", "ds", "uv"]
+    "http": ["ccf-api", "crypto", "ds"],
+    "http_client": ["ccf-api", "ds", "uv"]
   }
 }
