+++ src/node/jwt_key_auto_refresh.h
@@ -6,8 +6,8 @@
 #include "ccf/ds/nonstd.h"
 #include "ccf/service/tables/jwt.h"
 #include "enclave/http_rpc_context.h"
-#include "http/curl.h"
 #include "http/http_builder.h"
+#include "http_client/curl.h"
 #include "node/rpc/node_frontend.h"
 #include "tasks/basic_task.h"
 #include "tasks/task_system.h"
