+++ src/node/node_state.h
@@ -30,8 +30,8 @@
 #include "enclave/rpc_sessions.h"
 #include "encryptor.h"
 #include "history.h"
-#include "http/curl.h"
 #include "http/http_parser.h"
+#include "http_client/curl.h"
 #include "indexing/indexer.h"
 #include "js/global_class_ids.h"
 #include "network_state.h"
