+++ src/node/recovery_decision_protocol.cpp
@@ -9,7 +9,7 @@
 #include "ccf/service/tables/self_healing_open.h"
 #include "ccf/tx.h"
 #include "ccf/tx_id.h"
-#include "http/curl.h"
+#include "http_client/curl.h"
 #include "node_state.h"
 #include "tasks/basic_task.h"
 #include "tasks/task_system.h"
