+++ src/snapshots/fetch.h
@@ -5,8 +5,8 @@
 #include "ccf/ds/nonstd.h"
 #include "ccf/rest_verb.h"
 #include "ds/internal_logger.h"
-#include "http/curl.h"
 #include "http/http_builder.h"
+#include "http_client/curl.h"
 
 #include <charconv>
 #include <curl/curl.h>
